Clean api helper functions
